@@ -9,11 +9,11 @@
   "author": "Enrique Oriol",
   "license": "MIT",
   "devDependencies": {
-    "webpack": "^4.27.0",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.10"
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0"
   },
   "dependencies": {
-    "rxjs": "^6.4.0"
+    "rxjs": "^6.5.3"
   }
 }
